backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
email-validator==2.1.0.post1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
openai==1.3.0 